+++ static/index.html
@@ -21,7 +21,7 @@
     </script>
 </head>
 <body>
-    <script type="module" src="app.js"></script>
+    <script type="module" src="src/main.js"></script>
     <div id="root"></div>
 </body>
 </html>
